Adding back short ID generation changes (lost via merge)

+++ RustFlakes/UInt64Oxidation.cs
@@ -45,7 +45,7 @@ namespace RustFlakes
         {
             HandleTime();
 
-            return ((ulong) _lastOxidizedInMs << 32) + ((ulong) _identifier << 16) + _counter;
+            return (ulong)((_lastOxidizedInMs << 32) + (_identifier << 16) + _counter);
         }
 
         private void HandleTime()
